MCPAgent/pyproject.toml
[project]
name = "fastmcp-trial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.1",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.29.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-aws>=0.1.0",
    "langgraph>=0.1.0",
    "boto3>=1.28.0",
    "python-dotenv>=1.0.0",
]